+++ src/main/java/games/twinhead/morechests/client/MoreChestsClient.java
@@ -1,6 +1,7 @@
 package games.twinhead.morechests.client;
 
 import games.twinhead.morechests.MoreChests;
+import games.twinhead.morechests.client.screen.*;
 import games.twinhead.morechests.registry.ModBlockEntityRendererRegistry;
 import games.twinhead.morechests.registry.ScreenHandlerRegistry;
 import net.fabricmc.api.ClientModInitializer;