+++ src/main/java/games/twinhead/morechests/client/screen/GoldChestScreen.java
@@ -1,7 +1,8 @@
-package games.twinhead.morechests.client;
+package games.twinhead.morechests.client.screen;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import games.twinhead.morechests.screen.BasicChestScreenHandler;
+import games.twinhead.morechests.MoreChests;
+import games.twinhead.morechests.screen.GoldChestScreenHandler;
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import net.minecraft.client.render.GameRenderer;
 import net.minecraft.client.util.math.MatrixStack;
@@ -9,13 +10,28 @@ import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
-public class BasicChestScreen extends HandledScreen<BasicChestScreenHandler> {
+public class GoldChestScreen extends HandledScreen<GoldChestScreenHandler> {
+
+    public final int rows = 6;
+    public final int columns = 12;
 
     private static final Identifier TEXTURE =
-            new Identifier("minecraft", "textures/gui/container/generic_54.png");
+            new Identifier(MoreChests.MOD_ID, "textures/gui/container/generic_6x12.png");
 
-    public BasicChestScreen(BasicChestScreenHandler handler, PlayerInventory inventory, Text title) {
+    public GoldChestScreen(GoldChestScreenHandler handler, PlayerInventory inventory, Text title) {
         super(handler, inventory, title);
+        this.passEvents = false;
+        this.backgroundHeight = 113 + this.rows * 18;
+        this.playerInventoryTitleY = this.backgroundHeight - 93;
+        this.playerInventoryTitleX = this.backgroundWidth - 141;
+        this.backgroundWidth = 230;
+    }
+
+
+
+    @Override
+    protected void init() {
+        super.init();
     }
 
     @Override
@@ -26,8 +42,8 @@ public class BasicChestScreen extends HandledScreen<BasicChestScreenHandler> {
         int i = (width - backgroundWidth) / 2;
         int j = (height - backgroundHeight) / 2;
         //drawTexture(matrices, x, y, 0, 0, backgroundWidth, backgroundHeight);
-        this.drawTexture(matrices, i, j, 0, 0, this.backgroundWidth, 3 * 18 + 17);
-        this.drawTexture(matrices, i, j + 3 * 18 + 17, 0, 126, this.backgroundWidth, 94);
+        this.drawTexture(matrices, i, j, 0, 0, this.backgroundWidth, this.rows * 18 + 17);
+        this.drawTexture(matrices, i, j + this.rows * 18 + 17, 0, 126, this.backgroundWidth, 96);
     }
 
     @Override